Add files via upload
Point of sale in Python focus more on Cashiering System is a system that created in python programming language and this system is good for the beginners, this system has two features ordering and payment mode.

This cashiering system is easy to understand the codes that being provided and also easy to use for the people that beginners in the programming world and want to learn more in python programming. Also this system is way to help the students to improve their logic skills in programming.